refactor: extract shared settings dropdown component
