+++ components/dashboard/sidebar.tsx
@@ -22,10 +22,6 @@ import {
   Trophy,
   LogOut,
   Server,
-  Sun,
-  Moon,
-  Computer,
-  Globe,
   UserCog,
   TicketIcon,
   FileText,
@@ -39,17 +35,13 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
   DropdownMenuSeparator,
-  DropdownMenuSub,
-  DropdownMenuSubContent,
-  DropdownMenuSubTrigger,
 } from "@/components/ui/dropdown-menu";
 import { Button } from "@/components/ui/button";
 import { useTranslations } from "next-intl";
-import { useTheme } from "next-themes";
 import type { UserInfo } from "@/lib/api/types/auth";
 import { clashKingAssets } from "@/lib/theme";
 import { logout } from "@/lib/auth/logout";
-import { LANGUAGE_OPTIONS, LOCALE_MODE_COOKIE, getLocaleModeFromCookie, resolveBrowserLocale, type LocaleMode } from "@/lib/locale-preference";
+import { SettingsDropdown } from "@/components/settings-dropdown";
 
 interface SidebarProps {
   guildId: string;
@@ -69,7 +61,6 @@ export function Sidebar({ guildId, guildName, guildIcon, isLoading = false }: Si
 
   const [isDropdownOpen, setIsDropdownOpen] = useState(false);
   const [user, setUser] = useState<UserInfo | null>(null);
-  const [openSubmenu, setOpenSubmenu] = useState<string | null>(null);
 
   useEffect(() => {
     const storedUser = localStorage.getItem("user");
@@ -87,41 +78,6 @@ export function Sidebar({ guildId, guildName, guildIcon, isLoading = false }: Si
     setUser(null);
     router.push(`/${locale}`);
   };
-
-  const { theme, setTheme } = useTheme();
-  const [mounted, setMounted] = useState(false);
-  const [localeMode, setLocaleMode] = useState<LocaleMode>("manual");
-
-  useEffect(() => {
-    setMounted(true);
-    const currentMode = getLocaleModeFromCookie(document.cookie);
-    setLocaleMode(currentMode);
-
-    if (currentMode === "browser") {
-      const browserLocale = resolveBrowserLocale(navigator.languages);
-      if (browserLocale !== locale) {
-        document.cookie = `NEXT_LOCALE=${browserLocale}; path=/; max-age=31536000; SameSite=Lax`;
-        router.refresh();
-      }
-    }
-  }, [locale, router]);
-
-  const applyLocale = (newLocale: string, mode: LocaleMode) => {
-    // eslint-disable-next-line react-hooks/immutability
-    document.cookie = `${LOCALE_MODE_COOKIE}=${mode}; path=/; max-age=31536000; SameSite=Lax`;
-    // eslint-disable-next-line react-hooks/immutability
-    document.cookie = `NEXT_LOCALE=${newLocale}; path=/; max-age=31536000; SameSite=Lax`;
-    setLocaleMode(mode);
-    router.refresh();
-  };
-
-  const switchLocale = (newLocale: string) => applyLocale(newLocale, "manual");
-  const switchToBrowserLocale = () => applyLocale(resolveBrowserLocale(navigator.languages), "browser");
-
-  const currentLocale = mounted && localeMode === "browser"
-    ? resolveBrowserLocale(navigator.languages)
-    : locale;
-  const currentLanguage = LANGUAGE_OPTIONS.find((lang) => lang.code === currentLocale) ?? LANGUAGE_OPTIONS[0];
 
   const navigationSections = [
     {
@@ -389,105 +345,14 @@ export function Sidebar({ guildId, guildName, guildIcon, isLoading = false }: Si
       <div className="p-4 border-t border-border bg-background space-y-3">
         {/* Settings Button */}
         <div className="flex justify-center">
-          <DropdownMenu modal={false}>
-            <DropdownMenuTrigger asChild>
-              <Button variant="ghost" size="icon" className="border-border h-10 w-10">
-                <Settings className="h-5 w-5" />
-                <span className="sr-only">{tNav("settings")}</span>
-              </Button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent align="end" className="w-48 bg-popover border border-border shadow-2xl" sideOffset={4}>
-              {/* Theme Submenu */}
-              <DropdownMenuSub open={openSubmenu === "theme"} onOpenChange={(open) => setOpenSubmenu(open ? "theme" : null)}>
-                <DropdownMenuSubTrigger className="flex items-center space-x-2 hover:bg-accent/50 cursor-pointer">
-                  {mounted && theme === "dark" ? (
-                    <Moon className="h-4 w-4" />
-                  ) : theme === "light" ? (
-                    <Sun className="h-4 w-4" />
-                  ) : (
-                    <Computer className="h-4 w-4" />
-                  )}
-                  <span className="hover:text-primary">{tNav("theme")}</span>
-                </DropdownMenuSubTrigger>
-                <DropdownMenuSubContent className="bg-card border border-border shadow-2xl" sideOffset={2} alignOffset={-5}>
-                  <DropdownMenuItem
-                    onClick={() => setTheme("system")}
-                    className={`flex items-center space-x-2 hover:bg-accent/50 cursor-pointer ${
-                      theme === "system" ? "bg-primary/10 text-primary" : ""
-                    }`}
-                  >
-                    <Computer className="h-4 w-4" />
-                    <span className="hover:text-primary">{tNav("systemTheme")}</span>
-                  </DropdownMenuItem>
-                  <DropdownMenuItem
-                    onClick={() => setTheme("light")}
-                    className={`flex items-center space-x-2 hover:bg-accent/50 cursor-pointer ${
-                      theme === "light" ? "bg-primary/10 text-primary" : ""
-                    }`}
-                  >
-                    <Sun className="h-4 w-4" />
-                    <span className="hover:text-primary">{tNav("lightTheme")}</span>
-                  </DropdownMenuItem>
-                  <DropdownMenuItem
-                    onClick={() => setTheme("dark")}
-                    className={`flex items-center space-x-2 hover:bg-accent/50 cursor-pointer ${
-                      theme === "dark" ? "bg-primary/10 text-primary" : ""
-                    }`}
-                  >
-                    <Moon className="h-4 w-4" />
-                    <span className="hover:text-primary">{tNav("darkTheme")}</span>
-                  </DropdownMenuItem>
-                </DropdownMenuSubContent>
-              </DropdownMenuSub>
-
-              {/* Language Submenu */}
-              <DropdownMenuSub open={openSubmenu === "language"} onOpenChange={(open) => setOpenSubmenu(open ? "language" : null)}>
-                <DropdownMenuSubTrigger className="flex items-center space-x-2 hover:bg-accent/50 cursor-pointer">
-                  <div className="relative w-5 h-3.5 overflow-hidden rounded-sm border border-border/50">
-                    <Image
-                      src={`https://flagcdn.com/w40/${currentLanguage.flagCode}.png`}
-                      alt="Current language"
-                      fill
-                      sizes="20px"
-                      className="object-cover"
-                    />
-                  </div>
-                  <span className="hover:text-primary">{tNav("language")}</span>
-                </DropdownMenuSubTrigger>
-                <DropdownMenuSubContent className="bg-card border border-border shadow-2xl" sideOffset={2} alignOffset={-5}>
-                  <DropdownMenuItem
-                    onClick={switchToBrowserLocale}
-                    className={`flex items-center space-x-2 hover:bg-accent/50 cursor-pointer ${
-                      localeMode === "browser" ? "bg-primary/10 text-primary" : ""
-                    }`}
-                  >
-                    <Globe className="h-4 w-4" />
-                    <span className="hover:text-primary">{tNav("browserLanguage")}</span>
-                  </DropdownMenuItem>
-                  {LANGUAGE_OPTIONS.map((lang) => (
-                    <DropdownMenuItem
-                      key={lang.code}
-                      onClick={() => switchLocale(lang.code)}
-                      className={`flex items-center space-x-2 hover:bg-accent/50 cursor-pointer ${
-                        localeMode === "manual" && locale === lang.code ? "bg-primary/10 text-primary" : ""
-                      }`}
-                    >
-                      <div className="mr-2 relative w-5 h-3.5 overflow-hidden rounded-sm border border-border/50">
-                        <Image
-                          src={`https://flagcdn.com/w40/${lang.flagCode}.png`}
-                          alt={lang.name}
-                          fill
-                          sizes="20px"
-                          className="object-cover"
-                        />
-                      </div>
-                      <span className="hover:text-primary">{lang.name}</span>
-                    </DropdownMenuItem>
-                  ))}
-                </DropdownMenuSubContent>
-              </DropdownMenuSub>
-            </DropdownMenuContent>
-          </DropdownMenu>
+          <SettingsDropdown
+            locale={locale}
+            triggerButtonClassName="border-border h-10 w-10"
+            menuClassName="w-48 bg-popover border border-border shadow-2xl"
+            subTriggerClassName="flex items-center space-x-2 hover:bg-accent/50 cursor-pointer"
+            itemClassName="flex items-center space-x-2 hover:bg-accent/50 cursor-pointer"
+            textClassName="hover:text-primary"
+          />
         </div>
 
         {/* Branding */}
