+++ components/landing/navbar.tsx
@@ -4,17 +4,17 @@ import Link from "next/link";
 import Image from "next/image";
 import { useState, useEffect } from "react";
 import { useParams, useRouter } from "next/navigation";
-import { Menu, X, LogOut, ArrowRight, ChevronDown, Settings, Sun, Moon, Computer, Globe } from "lucide-react";
+import { Menu, X, LogOut, ArrowRight, ChevronDown } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger, DropdownMenuSub, DropdownMenuSubContent, DropdownMenuSubTrigger } from "@/components/ui/dropdown-menu";
+import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "@/components/ui/dropdown-menu";
 import { initiateDiscordLogin } from "@/lib/auth/discord-login";
 import { logout } from "@/lib/auth/logout";
 import { useTranslations } from "next-intl";
 import { useTheme } from "next-themes";
 import { clashKingAssets } from "@/lib/theme";
 import type { UserInfo } from "@/lib/api/types/auth";
-import { LANGUAGE_OPTIONS, LOCALE_MODE_COOKIE, getLocaleModeFromCookie, resolveBrowserLocale, type LocaleMode } from "@/lib/locale-preference";
+import { SettingsDropdown } from "@/components/settings-dropdown";
 
 export function Navbar() {
   const [isOpen, setIsOpen] = useState(false);
@@ -42,145 +42,6 @@ export function Navbar() {
     logout();
     setUser(null);
     router.push(`/${locale}`);
-  };
-
-  const SettingsDropdown = ({ align = "end" }: { align?: "start" | "end" }) => {
-    const { theme, setTheme } = useTheme();
-    const [mounted, setMounted] = useState(false);
-    const [openSubmenu, setOpenSubmenu] = useState<string | null>(null);
-    const [localeMode, setLocaleMode] = useState<LocaleMode>("manual");
-    const t = useTranslations("Navigation");
-
-    useEffect(() => {
-      setMounted(true);
-      const currentMode = getLocaleModeFromCookie(document.cookie);
-      setLocaleMode(currentMode);
-
-      if (currentMode === "browser") {
-        const browserLocale = resolveBrowserLocale(navigator.languages);
-        if (browserLocale !== locale) {
-          document.cookie = `NEXT_LOCALE=${browserLocale}; path=/; max-age=31536000; SameSite=Lax`;
-          router.refresh();
-        }
-      }
-    }, []);
-
-    const applyLocale = (newLocale: string, mode: LocaleMode) => {
-      document.cookie = `${LOCALE_MODE_COOKIE}=${mode}; path=/; max-age=31536000; SameSite=Lax`;
-      document.cookie = `NEXT_LOCALE=${newLocale}; path=/; max-age=31536000; SameSite=Lax`;
-      setLocaleMode(mode);
-      router.refresh();
-    };
-
-    const switchLocale = (newLocale: string) => applyLocale(newLocale, "manual");
-    const switchToBrowserLocale = () => applyLocale(resolveBrowserLocale(navigator.languages), "browser");
-
-    const currentLocale = mounted && localeMode === "browser"
-      ? resolveBrowserLocale(navigator.languages)
-      : locale;
-    const currentLanguage = LANGUAGE_OPTIONS.find((lang) => lang.code === currentLocale) ?? LANGUAGE_OPTIONS[0];
-
-    return (
-      <DropdownMenu modal={false}>
-        <DropdownMenuTrigger asChild>
-          <Button variant="ghost" size="icon" className="border-border h-12 w-12 md:h-9 md:w-9 [&_svg]:size-6 md:[&_svg]:size-5">
-            <Settings />
-            <span className="sr-only">{t("settings")}</span>
-          </Button>
-        </DropdownMenuTrigger>
-        <DropdownMenuContent align={align} className="w-48 bg-popover border border-border shadow-2xl" sideOffset={4} alignOffset={0}>
-          {/* Theme Submenu */}
-          <DropdownMenuSub open={openSubmenu === "theme"} onOpenChange={(open) => setOpenSubmenu(open ? "theme" : null)}>
-            <DropdownMenuSubTrigger className="flex items-center space-x-2 hover:!bg-transparent cursor-pointer">
-              {mounted && theme === "dark" ? (
-                <Moon className="h-4 w-4" />
-              ) : theme === "light" ? (
-                <Sun className="h-4 w-4" />
-              ) : (
-                <Computer className="h-4 w-4" />
-              )}
-              <span className="hover:!text-primary">{t("theme")}</span>
-            </DropdownMenuSubTrigger>
-            <DropdownMenuSubContent className="bg-card border border-border shadow-2xl" sideOffset={2} alignOffset={-5}>
-              <DropdownMenuItem
-                onClick={() => setTheme("system")}
-                className={`flex items-center space-x-2 hover:!bg-transparent cursor-pointer ${
-                  theme === "system" ? "bg-primary/10 text-primary" : ""
-                }`}
-              >
-                <Computer className="h-4 w-4" />
-                <span className="hover:!text-primary">{t("systemTheme")}</span>
-              </DropdownMenuItem>
-              <DropdownMenuItem
-                onClick={() => setTheme("light")}
-                className={`flex items-center space-x-2 hover:!bg-transparent cursor-pointer ${
-                  theme === "light" ? "bg-primary/10 text-primary" : ""
-                }`}
-              >
-                <Sun className="h-4 w-4" />
-                <span className="hover:!text-primary">{t("lightTheme")}</span>
-              </DropdownMenuItem>
-              <DropdownMenuItem
-                onClick={() => setTheme("dark")}
-                className={`flex items-center space-x-2 hover:!bg-transparent cursor-pointer ${
-                  theme === "dark" ? "bg-primary/10 text-primary" : ""
-                }`}
-              >
-                <Moon className="h-4 w-4" />
-                <span className="hover:!text-primary">{t("darkTheme")}</span>
-              </DropdownMenuItem>
-            </DropdownMenuSubContent>
-          </DropdownMenuSub>
-
-          {/* Language Submenu */}
-          <DropdownMenuSub open={openSubmenu === "language"} onOpenChange={(open) => setOpenSubmenu(open ? "language" : null)}>
-            <DropdownMenuSubTrigger className="flex items-center space-x-2 hover:!bg-transparent cursor-pointer">
-              <div className="relative w-5 h-3.5 overflow-hidden rounded-sm border border-border/50">
-                <Image
-                  src={`https://flagcdn.com/w40/${currentLanguage.flagCode}.png`}
-                  alt="Current language"
-                  fill
-                  sizes="20px"
-                  className="object-cover"
-                />
-              </div>
-              <span className="hover:!text-primary">{t("language")}</span>
-            </DropdownMenuSubTrigger>
-            <DropdownMenuSubContent className="bg-card border border-border shadow-2xl" sideOffset={2} alignOffset={-5}>
-              <DropdownMenuItem
-                onClick={switchToBrowserLocale}
-                className={`flex items-center space-x-2 hover:!bg-transparent cursor-pointer ${
-                  localeMode === "browser" ? "bg-primary/10 text-primary" : ""
-                }`}
-              >
-                <Globe className="h-4 w-4" />
-                <span className="hover:!text-primary">{t("browserLanguage")}</span>
-              </DropdownMenuItem>
-              {LANGUAGE_OPTIONS.map((lang) => (
-                <DropdownMenuItem
-                  key={lang.code}
-                  onClick={() => switchLocale(lang.code)}
-                  className={`flex items-center space-x-2 hover:!bg-transparent cursor-pointer ${
-                    localeMode === "manual" && locale === lang.code ? "bg-primary/10 text-primary" : ""
-                  }`}
-                >
-                  <div className="mr-2 relative w-5 h-3.5 overflow-hidden rounded-sm border border-border/50">
-                    <Image
-                      src={`https://flagcdn.com/w40/${lang.flagCode}.png`}
-                      alt={lang.name}
-                      fill
-                      sizes="20px"
-                      className="object-cover"
-                    />
-                  </div>
-                  <span className="hover:!text-primary">{lang.name}</span>
-                </DropdownMenuItem>
-              ))}
-            </DropdownMenuSubContent>
-          </DropdownMenuSub>
-        </DropdownMenuContent>
-      </DropdownMenu>
-    );
   };
 
   const logoSrc = mounted && (theme === "light" || resolvedTheme === "light")
@@ -225,7 +86,14 @@ export function Navbar() {
 
           {/* CTA Buttons */}
           <div className="hidden md:flex items-center space-x-4">
-            <SettingsDropdown />
+            <SettingsDropdown
+              locale={locale}
+              triggerButtonClassName="border-border h-12 w-12 md:h-9 md:w-9 [&_svg]:size-6 md:[&_svg]:size-5"
+              menuClassName="w-48 bg-popover border border-border shadow-2xl"
+              subTriggerClassName="flex items-center space-x-2 hover:!bg-transparent cursor-pointer"
+              itemClassName="flex items-center space-x-2 hover:!bg-transparent cursor-pointer"
+              textClassName="hover:!text-primary"
+            />
             {user ? (
               <DropdownMenu modal={false}>
                 <DropdownMenuTrigger asChild>
@@ -292,7 +160,15 @@ export function Navbar() {
                 <div className="space-y-2">
                   <div className="flex items-stretch gap-2">
                     <div className="flex items-center justify-center bg-accent rounded-lg">
-                      <SettingsDropdown align="start" />
+                      <SettingsDropdown
+                        locale={locale}
+                        align="start"
+                        triggerButtonClassName="border-border h-12 w-12 md:h-9 md:w-9 [&_svg]:size-6 md:[&_svg]:size-5"
+                        menuClassName="w-48 bg-popover border border-border shadow-2xl"
+                        subTriggerClassName="flex items-center space-x-2 hover:!bg-transparent cursor-pointer"
+                        itemClassName="flex items-center space-x-2 hover:!bg-transparent cursor-pointer"
+                        textClassName="hover:!text-primary"
+                      />
                     </div>
                     <div className="flex items-center justify-between p-1 bg-accent rounded-lg flex-1 min-w-0">
                       <div className="flex items-center space-x-3 min-w-0">
@@ -322,7 +198,15 @@ export function Navbar() {
                 <div className="space-y-2">
                   <div className="flex items-stretch gap-2">
                     <div className="flex items-center justify-center bg-accent rounded-lg">
-                      <SettingsDropdown align="start" />
+                      <SettingsDropdown
+                        locale={locale}
+                        align="start"
+                        triggerButtonClassName="border-border h-12 w-12 md:h-9 md:w-9 [&_svg]:size-6 md:[&_svg]:size-5"
+                        menuClassName="w-48 bg-popover border border-border shadow-2xl"
+                        subTriggerClassName="flex items-center space-x-2 hover:!bg-transparent cursor-pointer"
+                        itemClassName="flex items-center space-x-2 hover:!bg-transparent cursor-pointer"
+                        textClassName="hover:!text-primary"
+                      />
                     </div>
                     <Button
                       onClick={() => initiateDiscordLogin(locale)}
